pyproject.toml
[project]
name = "python-mal-2025"
version = "0.1.0"
description = "Mal for Python-prosjekt med 'best-practice' i 2025"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "statsmodels>=0.14",
    "scipy>=1.11",
    "google-cloud-bigquery>=3.40.1",
    "db-dtypes>=1.3.0",
    "polars>=1.38.1",
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.3",
    "tabulate>=0.10.0",
]

[dependency-groups]
dev = [
    "prek>=0.3.1",
    "ruff>=0.15.0",
    "types-pyyaml>=6.0.12.20250915",
]
notebooks = [
    "jupyter>=1.1.1",
]
quarto = [
    "quarto-cli>=1.8.26",
]

[tool.uv]
exclude-newer = "1 week"
